Make bulk_free() go faster

+++ libc/cosmo.h
@@ -25,8 +25,8 @@ int cosmo_futex_wake(_COSMO_ATOMIC(int) *, int, char);
 int cosmo_futex_wait(_COSMO_ATOMIC(int) *, int, char, int,
                      const struct timespec *);
 
-errno_t cosmo_stack_alloc(size_t *, size_t *, void **) libcesque;
-errno_t cosmo_stack_free(void *, size_t, size_t) libcesque;
+errno_t cosmo_stack_alloc(unsigned *, unsigned *, void **) libcesque;
+errno_t cosmo_stack_free(void *, unsigned, unsigned) libcesque;
 void cosmo_stack_clear(void) libcesque;
 void cosmo_stack_setmaxstacks(int) libcesque;
 int cosmo_stack_getmaxstacks(void) libcesque;
